Extract Test Client and ConversationsProcessor
