+++ src/main/kotlin/me/bors/slack/share/SlackConversationsProcessor.kt
@@ -5,7 +5,7 @@ import java.util.concurrent.CountDownLatch
 import java.util.concurrent.Executors
 import java.util.concurrent.LinkedBlockingQueue
 
-class ConversationsProcessor(private val slackClient: SlackClient) {
+open class SlackConversationsProcessor(protected val slackClient: SlackClient) {
     fun getConversations(): List<SlackConversation> {
         /* Multi-User conversations require a lot of requests to receive members and form a readable conversation name
            Concurrent execution helps to reduce execution time up to 4x
@@ -44,25 +44,6 @@ class ConversationsProcessor(private val slackClient: SlackClient) {
         latch.await()
 
         pool.shutdown()
-
-        return result.sortedByDescending { it.priority }
-    }
-
-    fun getConversationsSingleThreaded(): List<SlackConversation> {
-        val result = mutableListOf<SlackConversation>()
-
-        result.addAll(
-            slackClient.getChannels(listOf(ConversationType.IM))
-                .map { SlackConversation(it.id, slackClient.getUserName(it.user), it.priority ?: 0.0) }
-        )
-        result.addAll(
-            slackClient.getChannels(listOf(ConversationType.MPIM))
-                .map { SlackConversation(it.id, slackClient.getMultiUserGroupName(it.id), it.priority ?: 0.0) }
-        )
-        result.addAll(
-            slackClient.getChannels(listOf(ConversationType.PRIVATE_CHANNEL, ConversationType.PUBLIC_CHANNEL))
-                .map { SlackConversation(it.id, it.nameNormalized, it.priority ?: 0.0) }
-        )
 
         return result.sortedByDescending { it.priority }
     }
