+++ src/test/kotlin/SlackTestClient.kt
@@ -1,0 +1,4 @@
+package me.bors.slack.share
+
+class SlackTestClient : SlackClient() {
+}
